Adds source for cross platform killable process to test
As a way to clean the minikube repo dir.
