+++ cmd/minikube/cmd/delete_test.go
@@ -227,7 +227,42 @@ func TestDeleteAllProfiles(t *testing.T) {
 // then tries to execute the tryKillOne function on it;
 // if after tryKillOne the process still exists, we consider it a failure
 func TestTryKillOne(t *testing.T) {
-	dontkillmeProc := exec.Command("go", "run", "tools/pleasedontkillme/main.go")
+
+	var pleasedontkillmeSource = []byte(`
+package main
+
+import (
+	"os"
+	"os/signal"
+	"syscall"
+)
+
+// This is used to unittest functions that kill processes,
+// in a cross-platform way.
+func main() {
+	ch := make(chan os.Signal, 1)
+	done := make(chan struct{})
+	defer close(ch)
+
+	signal.Notify(ch, syscall.SIGHUP)
+	defer signal.Stop(ch)
+
+	go func() {
+		<-ch
+		close(done)
+	}()
+
+	<-done
+}
+`)
+	td := t.TempDir()
+	tmpfile := filepath.Join(td, "pleasedontkillme.go")
+
+	if err := os.WriteFile(tmpfile, pleasedontkillmeSource, 0o700); err != nil {
+		t.Fatalf("copying source to %s", tmpfile)
+	}
+
+	dontkillmeProc := exec.Command("go", "run", tmpfile)
 	err := dontkillmeProc.Start()
 	if err != nil {
 		t.Fatalf("while execing child process")
